correccion sexto punto

+++ 6.cpp
@@ -6,7 +6,7 @@ int main (void)
   std::cout.precision(16);
   std::cout.setf(std::ios::scientific);
   long double sum=0;
-  for(int ii=1; ii<1000000; ii++)
+  for(double ii=1; ii<1000000; ii++)
     {
       sum+=1/ii;
       std::cout<<ii<<"\t"<<sum<<std::endl;
